[lldb] Move FormatterBytecode.{h,def} to include/ (#179355)
To allow other headers to reference types declared in `FormatterBytecode.h`.

(cherry picked from commit 5587993548a6d764334c12b46616aa48606f2d6f)
