+++ lldb/source/DataFormatters/FormatterBytecode.cpp
@@ -6,7 +6,7 @@
 //
 //===----------------------------------------------------------------------===//
 
-#include "FormatterBytecode.h"
+#include "lldb/DataFormatters/FormatterBytecode.h"
 #include "lldb/Utility/LLDBLog.h"
 #include "lldb/ValueObject/ValueObject.h"
 #include "lldb/ValueObject/ValueObjectConstResult.h"
@@ -26,7 +26,7 @@ std::string toString(FormatterBytecode::OpCodes op) {
     const char *s = MNEMONIC;                                                  \
     return s ? s : #NAME;                                                      \
   }
-#include "FormatterBytecode.def"
+#include "lldb/DataFormatters/FormatterBytecode.def"
 #undef DEFINE_OPCODE
   }
   return llvm::utostr(op);
@@ -37,7 +37,7 @@ std::string toString(FormatterBytecode::Selectors sel) {
 #define DEFINE_SELECTOR(ID, NAME)                                              \
   case ID:                                                                     \
     return "@" #NAME;
-#include "FormatterBytecode.def"
+#include "lldb/DataFormatters/FormatterBytecode.def"
 #undef DEFINE_SELECTOR
   }
   return "@" + llvm::utostr(sel);
@@ -48,7 +48,7 @@ std::string toString(FormatterBytecode::Signatures sig) {
 #define DEFINE_SIGNATURE(ID, NAME)                                             \
   case ID:                                                                     \
     return "@" #NAME;
-#include "FormatterBytecode.def"
+#include "lldb/DataFormatters/FormatterBytecode.def"
 #undef DEFINE_SIGNATURE
   }
   return llvm::utostr(sig);
