+++ lldb/unittests/DataFormatter/FormatterBytecodeTest.cpp
@@ -1,4 +1,4 @@
-#include "DataFormatters/FormatterBytecode.h"
+#include "lldb/DataFormatters/FormatterBytecode.h"
 #include "lldb/Utility/StreamString.h"
 
 #include "gtest/gtest.h"
